+++ httpie/config.py
@@ -84,7 +84,6 @@ class Config(BaseConfigDict):
     about = 'HTTPie configuration file'
 
     DEFAULTS = {
-        'implicit_content_type': 'json',
         'default_options': []
     }
 
@@ -93,5 +92,21 @@ class Config(BaseConfigDict):
         self.update(self.DEFAULTS)
         self.directory = directory
 
+    def load(self):
+        super(Config, self).load()
+        self._migrate_implicit_content_type()
+
     def _get_path(self):
         return os.path.join(self.directory, self.name + '.json')
+
+    def _migrate_implicit_content_type(self):
+        """Migrate the removed implicit_content_type config option"""
+        try:
+            implicit_content_type = self.pop('implicit_content_type')
+        except KeyError:
+            pass
+        else:
+            if implicit_content_type == 'form':
+                self['default_options'].insert(0, '--form')
+            self.save()
+            self.load()